Move tests to node core and use manual graphql files (#2342)
* Move tests to node core and use manual graphql files

* Fix test
