+++ packages/node-core/test/migration-schemas/test_14_1000.graphql
@@ -1,0 +1,8 @@
+type Transfer @entity {
+  id: ID!
+  amount: BigInt!
+  blockNumber: Int!
+  date: Date!
+  #  from: Account!
+  #  to: Account!
+}
